Create start_chromium.sh
script to set up chromium and run it this forces chromium to use a /tmp folder on a tmpfs drive so it does not beat on the sd card. 
When chromium starts it creates many files on the card and has a large cache. These will all be set up on /tmp which is mounted
as a tmpfs.