Reset script messes with the mpd.service file
This script copies the 'correct' mpd.service file then changes it to comment out the user name (mpd). With the current systemd and mpd, this is no longer needed. If this is done, mpd cannot write to the log files and will not start. It is commented out and should be removed when 0.6 is released.