Update resolved.conf
DNSSEC should work. If it does not this is a hack (setting DNSSEC=no in resolved.conf) to get the RPi to boot and set the correct time. If this is needed, it is because the time is not set properly at boot and all DNSSEC queries have the wrong time and are discarded causing the machine to not set the correct time at boot. This needs to be fixed by other means, please.