Removed double quotes from shairport config data rate
Shairport-sync.conf data rate is "auto" or an integer data rate. Since alsa is set at a data rate in settings, we pass this value to the /etc/shairport-sync.conf file and there was a typo here which added the double quotes to the data rate e.g. "44100" and it needs to be bare. shairport-sync complains in the journal and rest it to 'auto' so it works, but I do not believe this was the intention.